[refactor] moved prompts.py into data folder and ran the unit tests and everything works
